+++ lib/multiply_2_matrix/multiply_2_matrix.h
@@ -1,7 +1,7 @@
 #ifndef _MULTIPLY_2_MATRIX_H_
 #define _MULTIPLY_2_MATRIX_H_
 
-double** multiply_2_matrix(double** matrix_1, double** matrix_2,
+float** multiply_2_matrix(float** matrix_1, float** matrix_2,
                            int num_row_matrix_1, int num_col_matrix_1,
                            int num_row_matrix_2, int num_col_matrix_2);
